1) Moved to the built in argument parser instead of my ugly custom one
2) Changed IUPAC aligner to get an object with functions for scoring
purposes instead of constant numbers.

3) Added a sequence motif feature. Forces high penalty for deletion or
insertions within a sequence motifs. Those are marked with lower case
sequence constraints.

4) Added a feature to remove penalty for insertion or removal of bulges
and internal loops smaller then a given size. This isn't ideal since if
we compare a stem to stem separated by a small buldge we still get a
(Stem (bulge (Stem (...))) vs (Stem (...)) comparison. The stem in the
second tree will be matched to one of the stems of the first tree
leaving the penalty for the other intact. For this reason the feature
will not be applied to the web-server.

5) Added gui options for generating image of a result and exporting as
tab separated values file.

6) varna location file is now set by config file

7) README is updated to new argument format and includes better
instruction for running the GUI client

8) Version was set to 2.0